feat: 新增曲线图

+++ src/Global/Types.h
@@ -5,7 +5,7 @@
  * Created Date: 2023-07-20 01:39:22
  * Author: 难为水
  * -----
- * Last Modified: 2023-08-08 07:50:32
+ * Last Modified: 2023-09-01 22:57:54
  * Modified By: 难为水
  * -----
  * CHANGELOG:
@@ -24,7 +24,7 @@ using BPInt_t = int;
 
 using PctFloat_t = double;
 
-using Value_t = int;
+using Value_t = long long;
 
 using Cof_t = double;
